# Fix performance issue in VF2 matching with unconnected graphs