--- a/networkx/algorithms/isomorphism/isomorphvf2.py
+++ b/networkx/algorithms/isomorphism/isomorphvf2.py
@@ -682,8 +682,8 @@ class DiGraphMatcher(GraphMatcher):
             # If all terminal sets are empty...
             # P(s) = (N_1 - M_1) x {min (N_2 - M_2)}
 
-            # elif not (T1_in or T2_in):   # as suggested by  [2], incorrect
-            else:  # as inferred from [1], correct
+            # else:  # as inferred from [1], correct
+            elif not (T2_out or T2_in):
                 node_2 = min(G2_nodes - set(self.core_2), key=min_key)
                 for node_1 in G1_nodes:
                     if node_1 not in self.core_1: