We have a use case where we need to search for small subgraph SG over a set of very large graphs G1, ..., Gk. We apply some pre-processing in order to cut down the search time by focusing on a small subgraph view of G1, ..., Gk which is guaranteed to contain all the occurrences of SG in G1, ..., Gk.

We found a scenario where a significant size reduction from close to 500k nodes to around 12k resulted in an increase in runtime instead of a reduction. The runtime with 500k nodes was around 2.13s while with the 12k nodes was 10.68s. The problem was with the scenario where both the `T1_out and T2_out` and `T1_in and T2_in` checks return `False` for unconnected graphs, in which case the `DiGraphMatcher` traverses all nodes in G1, ..., Gk.

However, this full traversal is only necessary if both `T2_out` and `T2_in` are empty. Lets assume, without loss of generality, that `T2_out` is not empty and that v is a node of G2 (SG in this case) in `T2_out`. We have that `T1_out` is also empty because otherwise the `T1_out and T2_out` check would have returned `True`. Therefore, v will never be included in the current partial mapping because no node of G1 can be mapped to v, otherwise that node would have been included in `T1_out`. Consequently, the current partial mapping isn't feasible and thus it's unnecessary to check any further candidate pairs.

This performance fix resulted in a runtime reduction from 10.68s to 0.33s for the 12k nodes subgraph views. The equivalent fix is also implemented for the `GraphMatcher`.